+++ packages/Datadog.Unity/Editor/DatadogConfigurationWindow.cs
@@ -39,54 +39,75 @@ namespace Datadog.Unity.Editor
             GUILayout.Label("SDK Options", EditorStyles.boldLabel);
 
             _options.Enabled = EditorGUILayout.ToggleLeft(
-                new GUIContent("Enable Datadog", "Whether the Datadog Plugin should be enabled or not."),
+                new GUIContent("Enable Datadog", DatadogHelpStrings.EnabledTooltip),
                 _options.Enabled);
             _options.OutputSymbols = EditorGUILayout.ToggleLeft(
                 new GUIContent(
                     "Output Symbol Files",
-                    "Whether the Datadog Plugin should output symbol files for crash reporting."),
+                    DatadogHelpStrings.OutputSymbolsTooltip),
                 _options.OutputSymbols);
 
-            _options.ClientToken = EditorGUILayout.TextField("Client Token", _options.ClientToken);
-            _options.Env = EditorGUILayout.TextField("Env", _options.Env);
-            _options.ServiceName = EditorGUILayout.TextField("Service Name", _options.ServiceName);
-            _options.Site = (DatadogSite)EditorGUILayout.EnumPopup("Datadog Site", _options.Site);
-            _options.BatchSize = (BatchSize)EditorGUILayout.EnumPopup("Batch Size", _options.BatchSize);
-            _options.UploadFrequency = (UploadFrequency)EditorGUILayout.EnumPopup("Upload Frequency", _options.UploadFrequency);
-            _options.BatchProcessingLevel = (BatchProcessingLevel)EditorGUILayout.EnumPopup("Batch Processing Level", _options.BatchProcessingLevel);
+            _options.ClientToken = EditorGUILayout.TextField(new GUIContent("Client Token", DatadogHelpStrings.ClientTokenTooltip), _options.ClientToken);
+            _options.Env = EditorGUILayout.TextField(new GUIContent("Env", DatadogHelpStrings.EnvTooltip), _options.Env);
+            _options.ServiceName = EditorGUILayout.TextField(new GUIContent("Service Name", DatadogHelpStrings.ServiceNameTooltip), _options.ServiceName);
+            _options.Site = (DatadogSite)EditorGUILayout.EnumPopup(new GUIContent("Datadog Site", DatadogHelpStrings.SiteTooltip), _options.Site);
+            _options.BatchSize = (BatchSize)EditorGUILayout.EnumPopup(new GUIContent("Batch Size", DatadogHelpStrings.BatchSizeTooltip), _options.BatchSize);
+            _options.UploadFrequency = (UploadFrequency)EditorGUILayout.EnumPopup(
+                new GUIContent("Upload Frequency", DatadogHelpStrings.UploadFrequencyTooltip), _options.UploadFrequency);
+            _options.BatchProcessingLevel = (BatchProcessingLevel)EditorGUILayout.EnumPopup(
+                new GUIContent("Batch Processing Level", DatadogHelpStrings.BatchProcessingLevelTooltip), _options.BatchProcessingLevel);
             _options.CrashReportingEnabled = EditorGUILayout.ToggleLeft(
-                new GUIContent("Enable Crash Reporting", "Whether to report native crashes to Datadog."),
+                new GUIContent("Enable Crash Reporting", DatadogHelpStrings.CrashReportingEnabledTooltip),
                 _options.CrashReportingEnabled);
 
             EditorGUILayout.Space();
             GUILayout.Label("Logging", EditorStyles.boldLabel);
             _options.ForwardUnityLogs = EditorGUILayout.ToggleLeft(
-                new GUIContent("Forward Unity Logs", "Whether calls to Debug.Log functions should be forwarded to Datadog."),
+                new GUIContent("Forward Unity Logs", DatadogHelpStrings.ForwardUnityLogsTooltip),
                 _options.ForwardUnityLogs);
-            _options.RemoteLogThreshold = (LogType)EditorGUILayout.EnumPopup("Remote Log Threshold", _options.RemoteLogThreshold);
+            _options.RemoteLogThreshold = (LogType)EditorGUILayout.EnumPopup(
+                new GUIContent("Remote Log Threshold", DatadogHelpStrings.RemoteLogThresholdTooltip), _options.RemoteLogThreshold);
 
             EditorGUILayout.Space();
             GUILayout.Label("RUM Options", EditorStyles.boldLabel);
             _options.RumEnabled = EditorGUILayout.ToggleLeft(
-                new GUIContent("Enable RUM", "Whether to enable Real User Monitoring (RUM)"),
+                new GUIContent("Enable RUM", DatadogHelpStrings.EnableRumTooltip),
                 _options.RumEnabled);
             EditorGUI.BeginDisabledGroup(!_options.RumEnabled);
+            _options.RumApplicationId = EditorGUILayout.TextField(
+                new GUIContent("RUM Application Id",DatadogHelpStrings.RUMApplicationIdTooltip), _options.RumApplicationId);
             _options.AutomaticSceneTracking = EditorGUILayout.ToggleLeft(
-                new GUIContent("Enable Automatic Scene Tracking", "Automatically start Datadog Views when Unity Scenes change"),
+                new GUIContent("Enable Automatic Scene Tracking", DatadogHelpStrings.EnableSceneTrackingTooltip),
                 _options.AutomaticSceneTracking);
-            _options.RumApplicationId = EditorGUILayout.TextField("RUM Application Id", _options.RumApplicationId);
-            _options.SessionSampleRate =
-                EditorGUILayout.FloatField("Session Sample Rate", _options.SessionSampleRate);
+            _options.SessionSampleRate = EditorGUILayout.FloatField(
+                new GUIContent("Session Sample Rate", DatadogHelpStrings.SessionSampleRateTooltip),
+                _options.SessionSampleRate);
             _options.SessionSampleRate = Math.Clamp(_options.SessionSampleRate, 0.0f, 100.0f);
-            _options.TraceSampleRate =
-                EditorGUILayout.FloatField("Trace Sample Rate", _options.TraceSampleRate);
-            _options.TraceContextInjection =
-                (TraceContextInjection)EditorGUILayout.EnumPopup("Trace Context Injection", _options.TraceContextInjection);
+            _options.TraceSampleRate = EditorGUILayout.FloatField(
+                new GUIContent("Trace Sample Rate", DatadogHelpStrings.TraceSampleRateTooltip),
+                _options.TraceSampleRate);
+            _options.TraceContextInjection = (TraceContextInjection)EditorGUILayout.EnumPopup(
+                new GUIContent("Trace Context Injection", DatadogHelpStrings.TraceContextInjectionTooltip),
+                _options.TraceContextInjection);
             _options.TraceSampleRate = Math.Clamp(_options.TraceSampleRate, 0.0f, 100.0f);
+            _options.TrackNonFatalAnrs = (NonFatalAnrDetectionOption)EditorGUILayout.EnumPopup(
+                new GUIContent("Track Non-Fatal ANRs", DatadogHelpStrings.TrackNonFatalAnrsTooltip),
+                _options.TrackNonFatalAnrs);
+            EditorGUILayout.BeginHorizontal();
+            _options.TrackNonFatalAppHangs = EditorGUILayout.ToggleLeft(
+                new GUIContent("Track Non-Fatal App Hangs", DatadogHelpStrings.TrackNonFatalAppHangsTooltip),
+                _options.TrackNonFatalAppHangs);
+            EditorGUI.BeginDisabledGroup(!_options.TrackNonFatalAppHangs);
+            _options.NonFatalAppHangThreshold = EditorGUILayout.FloatField(
+                new GUIContent("Threshold", DatadogHelpStrings.NonFatalAppHangThresholdTooltip),
+                _options.NonFatalAppHangThreshold);
+            EditorGUI.EndDisabledGroup();
+            GUILayout.FlexibleSpace();
+            EditorGUILayout.EndHorizontal();
 
             GUILayout.Space(12.0f);
 
-            GUILayout.Label("First Party Hosts", EditorStyles.boldLabel);
+            GUILayout.Label(new GUIContent("First Party Hosts", DatadogHelpStrings.FirstPartyHostsTooltip), EditorStyles.boldLabel);
             int toRemove = -1;
             for (int i = 0; i < _options.FirstPartyHosts.Count; ++i)
             {
@@ -118,10 +139,12 @@ namespace Datadog.Unity.Editor
             _showAdvancedOptions = EditorGUILayout.BeginFoldoutHeaderGroup(_showAdvancedOptions, "Advanced RUM Options");
             if (_showAdvancedOptions)
             {
-                _options.CustomEndpoint = EditorGUILayout.TextField("Custom Endpoint", _options.CustomEndpoint);
-                _options.SdkVerbosity = (CoreLoggerLevel)EditorGUILayout.EnumPopup("SDK Verbosity", _options.SdkVerbosity);
-                _options.TelemetrySampleRate =
-                    EditorGUILayout.FloatField("Telemetry Sample Rate", _options.TelemetrySampleRate);
+                _options.CustomEndpoint = EditorGUILayout.TextField(
+                    new GUIContent("Custom Endpoint", DatadogHelpStrings.CustomEndpointTooltip), _options.CustomEndpoint);
+                _options.SdkVerbosity = (CoreLoggerLevel)EditorGUILayout.EnumPopup(
+                    new GUIContent("SDK Verbosity", DatadogHelpStrings.SdkVerbosityTooltip), _options.SdkVerbosity);
+                _options.TelemetrySampleRate = EditorGUILayout.FloatField(
+                    new GUIContent("Telemetry Sample Rate", DatadogHelpStrings.TelemetrySampleRateTooltip), _options.TelemetrySampleRate);
                 _options.TelemetrySampleRate = Math.Clamp(_options.TelemetrySampleRate, 0.0f, 100.0f);
             }
             EditorGUILayout.EndFoldoutHeaderGroup();