+++ packages/Datadog.Unity/Editor/iOS/PostBuildProcess.cs
@@ -4,6 +4,7 @@
 #if UNITY_EDITOR_OSX
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Text;
@@ -176,11 +177,15 @@ func initializeDatadog() {{
 
                 sb.AppendLine($"    rumConfig.sessionSampleRate = {options.SessionSampleRate}");
                 sb.AppendLine($"    rumConfig.telemetrySampleRate = {options.TelemetrySampleRate}");
+                var appHangThreshold =
+                    options.TrackNonFatalAppHangs ? options.NonFatalAppHangThreshold.ToString(CultureInfo.InvariantCulture) : "nil";
+                sb.AppendLine(
+                    $"    rumConfig.appHangThreshold = {appHangThreshold}");
 
                 // Uncomment to enable RUM Configuration Telemetry
-    //             sb.AppendLine(@"    rumConfig._internal_mutation {
-    //     $0.configurationTelemetrySampleRate = 100.0
-    // }");
+                //             sb.AppendLine(@"    rumConfig._internal_mutation {
+                //     $0.configurationTelemetrySampleRate = 100.0
+                // }");
                 sb.AppendLine("    RUM.enable(with: rumConfig)");
             }