+++ packages/Datadog.Unity/Tests/Editor/DatadogConfigurationOptionsTests.cs
@@ -70,6 +70,9 @@ namespace Datadog.Unity.Editor.Tests
             Assert.AreEqual(20.0f, options.TraceSampleRate);
             Assert.AreEqual(TraceContextInjection.All, options.TraceContextInjection);
             Assert.AreEqual(20.0f, options.TelemetrySampleRate);
+            Assert.AreEqual(NonFatalAnrDetectionOption.Disabled, options.TrackNonFatalAnrs);
+            Assert.AreEqual(false, options.TrackNonFatalAppHangs);
+            Assert.AreEqual(0.25f, options.NonFatalAppHangThreshold);
         }
     }
 }